backend/requirements.txt
Django
django-extensions
django-environ
psycopg2-binary
black
pylint
pylint-django
pytest-django
Faker
factory-boy
django-cors-headers
djangorestframework
django-filter
djangorestframework-simplejwt
langchain
langchain-google-genai
google-generativeai
toons